+++ ui/litellm-dashboard/src/app/(dashboard)/usage/_components/components/EntityUsage/EntityUsage.test.tsx
@@ -39,11 +39,21 @@ vi.mock("../EndpointUsage/EndpointUsage", () => ({
 }));
 
 vi.mock("@/components/UsagePage/components/EntityUsage/TopKeyView", () => ({
-  default: () => <div>Top Keys</div>,
+  default: ({ topKeys }: { topKeys: { api_key: string; spend: number }[] }) => (
+    <div>
+      <span>Top Keys</span>
+      <span>{`top-keys:${topKeys.map((row) => `${row.api_key}=${row.spend}`).join("|")}`}</span>
+    </div>
+  ),
 }));
 
 vi.mock("./TopModelView", () => ({
-  default: () => <div>Top Models</div>,
+  default: ({ topModels }: { topModels: { key: string; spend: number }[] }) => (
+    <div>
+      <span>Top Models</span>
+      <span>{`top-models:${topModels.map((row) => `${row.key}=${row.spend}`).join("|")}`}</span>
+    </div>
+  ),
 }));
 
 vi.mock("@/components/EntityUsageExport/EntityUsageExportModal", () => ({
@@ -921,5 +931,40 @@ describe("EntityUsage", () => {
     });
     expect(screen.queryByAltText("zzz-internal logo")).not.toBeInTheDocument();
     expect(screen.getByText("z")).toBeInTheDocument();
+  });
+
+  it("feeds the key, model and agent tables from their own breakdowns", async () => {
+    const usageMetrics = {
+      spend: 30.75,
+      api_requests: 300,
+      successful_requests: 290,
+      failed_requests: 10,
+      total_tokens: 15000,
+      prompt_tokens: 9000,
+      completion_tokens: 6000,
+      cache_read_input_tokens: 0,
+      cache_creation_input_tokens: 0,
+    };
+    mockTeamDailyActivityCall.mockResolvedValue({
+      ...mockSpendData,
+      results: [
+        {
+          ...mockSpendData.results[0],
+          breakdown: {
+            ...mockSpendData.results[0].breakdown,
+            model_groups: { "gpt-4o": { metrics: { ...usageMetrics, spend: 70.25 }, metadata: {} } },
+            api_keys: { "sk-abc": { metrics: usageMetrics, metadata: { key_alias: "prod-key", team_id: null } } },
+          },
+        },
+      ],
+    });
+
+    render(<EntityUsage {...defaultProps} entityType="team" />);
+
+    await waitFor(() => {
+      expect(screen.getByText("top-keys:sk-abc=30.75")).toBeInTheDocument();
+    });
+    expect(screen.getByText("top-models:gpt-4o=70.25")).toBeInTheDocument();
+    expect(screen.getByText(/^top-models:Code Review Agent=/)).toBeInTheDocument();
   });
 });
